R project started
Using Github for first time. Github created local repository and then I
opened that folder as an R project. I guess that RStudio will use the
local repository as the place to store everything and then those changes
will be synchronized with git (instead of Dropbox, as I used to do)